# Fix 6052, do not recreate bundle class loader which break cache system