## Related Issues

- closes #6052 Performance: Very sluggish page load when languages and all facets enabled in root dv.

Before:

```
> for i in $(seq 10);do time curl -o /dev/null -s https://dataverse-test.ird.fr ;done
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 0% cpu 3,773 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 0% cpu 3,614 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 0% cpu 3,820 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,02s system 1% cpu 3,517 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 0% cpu 3,580 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 0% cpu 4,453 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 0% cpu 3,976 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 0% cpu 3,754 total
^C
```

After:

```
> for i in $(seq 10);do time curl -o /dev/null -s https://dataverse-test.ird.fr ;done
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,02s system 7% cpu 0,447 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 5% cpu 0,478 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 9% cpu 0,334 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 8% cpu 0,324 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,01s user 0,01s system 4% cpu 0,518 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 8% cpu 0,349 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 7% cpu 0,360 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 7% cpu 0,366 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 8% cpu 0,339 total
curl -o /dev/null -s https://dataverse-test.ird.fr  0,02s user 0,01s system 7% cpu 0,401 total
```

## Pull Request Checklist

- [ ] Unit [tests][] completed
- [ ] Integration [tests][]: None
- [ ] Deployment requirements, [SQL updates][], [Solr updates][], etc.: None
- [ ] [Documentation][docs] completed
- [ ] Merged latest from "develop" [branch][] and resolved conflicts

[tests]: http://guides.dataverse.org/en/latest/developers/testing.html
[SQL updates]: http://guides.dataverse.org/en/latest/developers/sql-upgrade-scripts.html
[Solr updates]: https://github.com/IQSS/dataverse/blob/develop/conf/solr/7.3.0/schema.xml
[docs]: http://guides.dataverse.org/en/latest/developers/documentation.html
[branch]: http://guides.dataverse.org/en/latest/developers/branching-strategy.html